# alertmanager/docs/swagger-v1.yml
...
swagger: '2.0'
info:
  title: Alertmanager Configurer Model Definitions and Paths
  description: Alertmanager Configurer REST APIs
  version: 1.0.2
basePath: /v1
paths:
...
  /tenants:
    get:
      summary: List configured tenants
      tags:
        - Tenants
      responses:
        '200':
          description: List of configured tenants
        default:
          $ref: '#/responses/UnexpectedError'
...
  /global:
    get:
      summary: Retrieve alertmanager global config
      tags:
        - Global
      responses:
        '200':
          description: Alerting tree
          schema:
            $ref: '#/definitions/global_config'
    post:
      summary: Modify alertmanager global config
      tags:
        - Global
      parameters:
        - in: body
          name: global_config
          description: Global config to be set
          required: true
          schema:
            $ref: '#/definitions/global_config'
      responses:
        '200':
          description: OK
        default:
          $ref: '#/responses/UnexpectedError'
...
definitions:
...
  global_config:
    type: object
    properties:
      resolve_timeout:
        type: string
      http_config:
        $ref: '#/definitions/http_config'
      smtp_from:
        type: string
      smtp_hello:
        type: string
      smtp_smarthost:
        type: string
      smtp_auth_username:
        type: string
      smtp_auth_password:
        type: string
      smtp_auth_secret:
        type: string
      smtp_auth_identity:
        type: string
      smtp_require_tls:
        type: string
      slack_api_url:
        type: string
      pagerduty_url:
        type: string
      hipchat_api_url:
        type: string
      hipchat_auth_token:
        type: string
      opsgenie_api_url:
        type: string
      opsgenie_api_key:
        type: string
      wechat_api_url:
        type: string
      wechat_api_corp_id:
        type: string
      victorops_api_url:
        type: string
      victorops_api_key:
        type: string
  http_config:
    type: object
    properties:
      basic_auth:
        $ref: '#/definitions/basic_auth'
      bearer_token:
        type: string
      proxy_url:
        type: string
      tls_config:
        $ref: '#/definitions/tls_config'
  basic_auth:
    type: object
    properties:
      username:
        type: string
      password:
        type: string
...
  tls_config:
    type: object
    properties:
      server_name:
        type: string
      insecure_skip_verify:
        type: boolean
...
  error:
    type: object
    required:
      - message
    properties:
      message:
        example: Error string
        type: string
responses:
  UnexpectedError:
    description: Unexpected Error
    schema:
      $ref: '#/definitions/error'